refactor: create media folder and move video inside
